requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

blinker==1.8.2 ; python_version >= "3.10"
certifi==2024.8.30 ; python_version >= "3.10"
charset-normalizer==3.4.0 ; python_version >= "3.10"
click==8.1.7 ; python_version >= "3.10"
colorama==0.4.6 ; python_version >= "3.10" and platform_system == "Windows"
contourpy==1.3.0 ; python_version >= "3.10"
cycler==0.12.1 ; python_version >= "3.10"
filelock==3.16.1 ; python_version >= "3.10"
flask==3.0.3 ; python_version >= "3.10"
fonttools==4.54.1 ; python_version >= "3.10"
fsspec==2024.10.0 ; python_version >= "3.10"
idna==3.10 ; python_version >= "3.10"
itsdangerous==2.2.0 ; python_version >= "3.10"
jinja2==3.1.4 ; python_version >= "3.10"
kiwisolver==1.4.7 ; python_version >= "3.10"
markupsafe==3.0.2 ; python_version >= "3.10"
matplotlib==3.9.2 ; python_version >= "3.10"
mpmath==1.3.0 ; python_version >= "3.10"
networkx==3.4.2 ; python_version >= "3.10"
numpy==2.1.2 ; python_version >= "3.10"
nvidia-cublas-cu12==12.1.3.1 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-cuda-cupti-cu12==12.1.105 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-cuda-nvrtc-cu12==12.1.105 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-cuda-runtime-cu12==12.1.105 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-cudnn-cu12==9.1.0.70 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-cufft-cu12==11.0.2.54 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-curand-cu12==10.3.2.106 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-cusolver-cu12==11.4.5.107 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-cusparse-cu12==12.1.0.106 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-nccl-cu12==2.21.5 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-nvjitlink-cu12==12.6.77 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-nvtx-cu12==12.1.105 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
opencv-python==4.10.0.84 ; python_version >= "3.10"
packaging==24.1 ; python_version >= "3.10"
pandas==2.2.3 ; python_version >= "3.10"
pillow==11.0.0 ; python_version >= "3.10"
psutil==6.1.0 ; python_version >= "3.10"
py-cpuinfo==9.0.0 ; python_version >= "3.10"
pyparsing==3.2.0 ; python_version >= "3.10"
python-dateutil==2.9.0.post0 ; python_version >= "3.10"
python-dotenv==1.0.1 ; python_version >= "3.10"
pytz==2024.2 ; python_version >= "3.10"
pyyaml==6.0.2 ; python_version >= "3.10"
requests==2.32.3 ; python_version >= "3.10"
scipy==1.14.1 ; python_version >= "3.10"
seaborn==0.13.2 ; python_version >= "3.10"
setuptools==75.2.0 ; python_version >= "3.12"
six==1.16.0 ; python_version >= "3.10"
sympy==1.13.1 ; python_version >= "3.10"
torch==2.5.0+cu121 ; python_version >= "3.10"
torchvision==0.20.0+cu121 ; python_version >= "3.10"
tqdm==4.66.5 ; python_version >= "3.10"
triton==3.1.0 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version < "3.13" and python_version >= "3.10"
typing-extensions==4.12.2 ; python_version >= "3.10"
tzdata==2024.2 ; python_version >= "3.10"
ultralytics-thop==2.0.9 ; python_version >= "3.10"
ultralytics==8.3.21 ; python_version >= "3.10"
urllib3==2.2.3 ; python_version >= "3.10"
werkzeug==3.0.4 ; python_version >= "3.10"
